Labirynt/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# LabirynthGen.py: 1
Pillow == 5.1.0
